refactor review editor into subcomponents
