+++ src/components/reviews/ReviewNotesTags.tsx
@@ -1,0 +1,108 @@
+import * as React from "react";
+import SectionLabel from "@/components/reviews/SectionLabel";
+import Input from "@/components/ui/primitives/Input";
+import IconButton from "@/components/ui/primitives/IconButton";
+import Textarea from "@/components/ui/primitives/Textarea";
+import { Tag, Plus } from "lucide-react";
+
+export default function ReviewNotesTags({
+  notes,
+  onNotesChange,
+  onNotesBlur,
+  tags,
+  onTagsChange,
+}: {
+  notes: string;
+  onNotesChange: (v: string) => void;
+  onNotesBlur: () => void;
+  tags: string[];
+  onTagsChange: (next: string[]) => void;
+}) {
+  const [draftTag, setDraftTag] = React.useState("");
+
+  function addTag(tagRaw: string) {
+    const t = tagRaw.trim().replace(/^#/, "");
+    if (!t || tags.includes(t)) return;
+    const next = [...tags, t];
+    onTagsChange(next);
+  }
+
+  function removeTag(t: string) {
+    const next = tags.filter((x) => x !== t);
+    onTagsChange(next);
+  }
+
+  return (
+    <>
+      {/* Tags */}
+      <div>
+        <SectionLabel>Tags</SectionLabel>
+        <div className="mt-1 flex items-center gap-2">
+          <div className="relative flex-1">
+            <Tag className="pointer-events-none absolute left-4 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-muted-foreground" />
+            <Input
+              value={draftTag}
+              onChange={(e) => setDraftTag(e.target.value)}
+              placeholder="Add tag and press Enter"
+              className="pl-6"
+              onKeyDown={(e) => {
+                if (e.key === "Enter") {
+                  e.preventDefault();
+                  addTag(draftTag);
+                  setDraftTag("");
+                }
+              }}
+            />
+          </div>
+
+          <IconButton
+            aria-label="Add tag"
+            title="Add tag"
+            size="md"
+            iconSize="sm"
+            variant="solid"
+            onClick={() => {
+              addTag(draftTag);
+              setDraftTag("");
+            }}
+          >
+            <Plus />
+          </IconButton>
+        </div>
+
+        {tags.length === 0 ? (
+          <div className="mt-2 text-sm text-muted-foreground/80">No tags yet.</div>
+        ) : (
+          <div className="mt-2 flex flex-wrap items-center gap-2">
+            {tags.map((t) => (
+              <button
+                key={t}
+                type="button"
+                className="chip h-9 px-4 text-sm group inline-flex items-center gap-1"
+                title="Remove tag"
+                onClick={() => removeTag(t)}
+              >
+                <span>#{t}</span>
+                <span className="opacity-0 transition-opacity group-hover:opacity-100">✕</span>
+              </button>
+            ))}
+          </div>
+        )}
+      </div>
+
+      {/* Notes */}
+      <div>
+        <SectionLabel>Notes</SectionLabel>
+        <Textarea
+          value={notes}
+          onChange={(e) => onNotesChange(e.target.value)}
+          onBlur={onNotesBlur}
+          placeholder="Key moments, mistakes to fix, drills to run…"
+          className="rounded-2xl"
+          resize="resize-y"
+          textareaClassName="min-h-[180px] leading-relaxed"
+        />
+      </div>
+    </>
+  );
+}
